+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 ## Unreleased
 ### Added
 - slug model / serializer mixins
+- extendible health check
 - no_button_center modal type
 - continue_center modal type
 ### Fixed
 - auth redirect in standalone auth mode
+- fixes issue with meta data cache expiry if expiry was set to never and api throttling was encountered
 
 
 ## 1.11.0